Support native browser dark theme
This approach allows the browser to override everything with its default light and dark stylesheet per the client's preferences.

Not supported in all browsers yet but certainly the simplest and most elegant solution.

Only niggle is that code snippets don't support it natively. I'll raise an issue for that. This commit is an improvement regardless.